pyproject.toml
[project]
name = "pyms-django-chassis"
version = "1.0.0"
description = "Open-source Django chassis for building microservices with cross-cutting concerns out of the box"
license = "MIT"
authors = [{ name = "PyMS Contributors" }]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Typing :: Typed",
]

dependencies = [
    "django>=4.2",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=3.14.0",
    "django-filter>=23.3",
    "pydantic>=2.8.2",
    "python-json-logger>=2.0.7",
    "toml>=0.10.2",
    "whitenoise>=6.4.0",
    "sqlparse>=0.5.0",
    "drf-extra-fields>=3.7.0",
]

[project.optional-dependencies]
monitoring = [
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-propagator-b3>=1.26.0",
    "opentelemetry-exporter-otlp-proto-http>=1.26.0",
]
aws = [
    "boto3>=1.34.162",
]
tenant = [
    "django-tenants>=3.4.8",
    "psycopg2-binary>=2.9.5",
]
docs = [
    "drf-spectacular>=0.25.1",
]
restql = [
    "django-restql>=0.15.3",
]
import-export = [
    "django-import-export>=3.3.6",
]
dev-tools = [
    "django-debug-toolbar>=3.8.1",
    "django-extensions>=4.1",
]
daas = [
    "pyms-django-chassis[tenant,docs,restql,import-export,monitoring,aws]",
]
baas = [
    "pyms-django-chassis[tenant,docs,restql,monitoring,aws]",
]
all = [
    "pyms-django-chassis[monitoring,aws,tenant,docs,restql,import-export,dev-tools]",
]

[project.scripts]
pyms-django = "pyms_django.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "pytest-django>=4.7.0",
    "pre-commit>=3.5.0",
    "mypy>=1.7.1",
    "ruff>=0.8.4",
    "bandit>=1.7.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyms_django"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "COM", "C4", "PT", "RET", "SIM", "TCH", "ARG", "PTH", "ERA"]
